+++ IVTC/Filtre_JPSDR_IVTC.cpp
@@ -600,7 +600,7 @@ protected:
 	MT_Data_Info MT_Data[MAX_MT_THREADS];
 	uint8_t threads_number,total_cpu;
 	bool threadpoolAllocated;
-	uint16_t UserId;
+	uint32_t UserId;
 
 	uint32_t histogramme_MT[MAX_MT_THREADS][256];
 	uint32_t repartition_MT[MAX_MT_THREADS][256];
