+++ Saturation/Filtre_JPSDR_Saturation.cpp
@@ -192,7 +192,7 @@ protected:
 	MT_Data_Info MT_Data[MAX_MT_THREADS];
 	uint8_t threads_number,total_cpu;
 	bool threadpoolAllocated;
-	uint16_t UserId;
+	uint32_t UserId;
 
 	ThreadPoolFunction StaticThreadpoolF;
